backend/pyproject.toml
[project]
name = "mirofish-backend"
version = "0.2.0"
description = "MiroFish Lite — Cost-optimized cloud multi-agent simulation (Gemini + Supabase)"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "MiroFish Team" }
]

dependencies = [
    # 核心框架
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    # LLM: Google Gemini 2.0 Flash (replaces OpenAI as primary LLM)
    "google-generativeai>=0.8.0",
    # Memory: Supabase PostgreSQL + pgvector (replaces Zep Cloud)
    "supabase>=2.0.0",
    # OASIS social simulation engine
    # Kept: simulation scripts route through Gemini's OpenAI-compatible endpoint
    "openai>=1.0.0",
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",
    # File processing
    "PyMuPDF>=1.24.0",
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "google-genai>=1.68.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
